+++ src/main/java/answers/TwoBitSplit.java
@@ -5,40 +5,40 @@ import answers.Match;
 
 public class TwoBitSplit {
 	//A array only matches with B so we need to differentiate them
-	private ArrayList<Byte> A_0 = new ArrayList<Byte>(); //0
-	private ArrayList<Byte> A_1 = new ArrayList<Byte>(); //1
+	private ArrayList<Integer> A_0 = new ArrayList<Integer>(); //0
+	private ArrayList<Integer> A_1 = new ArrayList<Integer>(); //1
 
 	//B array only matches with A so we need to differentiate them
-	private ArrayList<Byte> B_0 = new ArrayList<Byte>(); //0
-	private ArrayList<Byte> B_1= new ArrayList<Byte>(); //1
+	private ArrayList<Integer> B_0 = new ArrayList<Integer>(); //0
+	private ArrayList<Integer> B_1 = new ArrayList<Integer>(); //1
 	
 	private TwoBitSplit[] children;
 
 	private boolean childA, childB, bottom = false;
-	private byte depthA, depthB = 0;
-	private byte maxDepth, depth;
+	private int depthA, depthB = 0;
+	private int maxDepth, depth;
 
 
-	public TwoBitSplit(byte depth) {
+	public TwoBitSplit(int depth) {
 		this.depth = depth;
 		children = new TwoBitSplit[2];
 	}
-	public byte split(int[] portfolios, ArrayList<Byte> indices_A, ArrayList<Byte> indices_B) {
+	public int split(int[] portfolios, ArrayList<Integer> indices_A, ArrayList<Integer> indices_B) {
 		//Initialise shift
-		byte shift = (byte) (Question1.NUMBER_OF_BITS - (Question1.BITS_OBSERVED*(depth+1)));
+		int shift = (int) (Question1.NUMBER_OF_BITS - (Question1.BITS_OBSERVED*(depth+1)));
 		//Loop through the integers corresponding to indices_A 
-		for(byte i=0; i<indices_A.size(); i++) {
+		for(int i=0; i<indices_A.size(); i++) {
 			//Take first 2 elements and place into the 4 arrays to pass into child node
-			byte indexA = indices_A.get(i);
+			int indexA = indices_A.get(i);
 			boolean firstTwoBits = (portfolios[indexA]>>shift&1)==1;
 			//Either a 1 or is not a 1 (a 0)
 			if(firstTwoBits) { A_1.add(indexA); }
 			else { A_0.add(indexA); }
 		}
 		//Loop through the integers corresponding to indices_B
-		for(byte j=0; j<indices_B.size(); j++) {
+		for(int j=0; j<indices_B.size(); j++) {
 			//Take first 2 elements and place into the 4 arrays to pass into child node
-			byte indexB = indices_B.get(j);
+			int indexB = indices_B.get(j);
 			boolean firstTwoBits = (portfolios[indexB]>>shift&1)==1;
 			//Either a 1 or is not a 1 (a 0)
 			if(firstTwoBits) { B_1.add(indexB); }
@@ -47,8 +47,8 @@ public class TwoBitSplit {
 
 		//Since the inputs were passed such that A and B match
 		//If we find matches BETWEEN A and B they are further matches i.e
-		if(A_0.size()>0 && B_1.size()>0) { children[0] = new TwoBitSplit((byte) (this.depth+1)); childA = true; } 
-		if(A_1.size()>0 && B_0.size()>0) { children[1] = new TwoBitSplit((byte) (this.depth+1)); childB = true; }
+		if(A_0.size()>0 && B_1.size()>0) { children[0] = new TwoBitSplit((int) (this.depth+1)); childA = true; } 
+		if(A_1.size()>0 && B_0.size()>0) { children[1] = new TwoBitSplit((int) (this.depth+1)); childB = true; }
 
 		//If no children we are at the end of our search
 		bottom = (!childA && !childB) || (depth==16);
@@ -68,7 +68,7 @@ public class TwoBitSplit {
 	public int getDepth() {
 		return maxDepth;
 	}
-	public ArrayList<Match> getPossibleMatches(int []portfolios, ArrayList<Byte> indexesA, ArrayList<Byte> indexesB) {
+	public ArrayList<Match> getPossibleMatches(int []portfolios, ArrayList<Integer> indexesA, ArrayList<Integer> indexesB) {
 		ArrayList<Match> solutions = new ArrayList<Match>();
 
 		if(bottom) {
@@ -99,16 +99,16 @@ public class TwoBitSplit {
 		return solutions;
 	}
 	public ArrayList<Match> getNextConsecutiveRun(int[] portfolios, ArrayList<Match> solutions, 
-			ArrayList<Byte> indexesA, ArrayList<Byte> indexesB) {
+			ArrayList<Integer> indexesA, ArrayList<Integer> indexesB) {
 		//This function will skip out one bit in order to find the next consecutive run of bits 
 		//Applies to the case where very many of the portfolios give the same string of 1s
 		boolean foundArray = false;
 		//skipping one bit which we know has no more consecutive but this strand must be best...
-		byte mSkipTo = (byte)(depth + 1);
+		int mSkipTo = (int)(depth + 1);
 		
 		TwoBitSplit skipChild; 
 		
-		byte skipDepth = 0;
+		int skipDepth = 0;
 		ArrayList<Match> newSolutions = new ArrayList<Match>();
 		
 		while(!foundArray) {
@@ -116,7 +116,7 @@ public class TwoBitSplit {
 		
 			skipDepth = skipChild.split(portfolios, indexesA, indexesB);
 			
-			if(skipDepth!=(byte)mSkipTo) {
+			if(skipDepth!=(int)mSkipTo) {
 				//We managed to find a better match
 				//otherwise collate all other results
 				newSolutions.addAll(skipChild.getPossibleMatches(portfolios, indexesA, indexesB));
